+++ lib/resources/epic_services.dart
@@ -1,17 +1,33 @@
+/// epic service routes object
 class EpicServices {
+  /// epic account service
   String get accountService =>
       "https://account-public-service-prod03.ol.epicgames.com";
+
+  /// launcher caldera service
   String get calderaService =>
       "https://caldera-service-prod.ecosec.on.epicgames.com";
+
+  /// br stats service
   String get statsService =>
       "https://statsproxy-public-service-live.ol.epicgames.com";
+
+  /// lightswitch service for epic service status
   String get lightswitchService =>
       "https://lightswitch-public-service-prod06.ol.epicgames.com";
+
+  /// fortnite service
   String get fortniteService =>
       "https://fortnite-public-service-prod11.ol.epicgames.com";
+
+  /// fortnite party service
   String get partyService => "https://party-service-prod.ol.epicgames.com";
+
+  /// user search service
   String get userSearchService =>
       "https://user-search-service-prod.ol.epicgames.com";
+
+  /// user friends service
   String get friendService =>
       "https://friends-public-service-prod.ol.epicgames.com";
 }